+++ source/main/zettlr-ipc.js
@@ -135,10 +135,6 @@ class ZettlrIPC
             this._app.saveFile(cnt);
             break;
 
-            case 'file-revert':
-            this._app.revert();
-            break;
-
             case 'dir-open':
             // Client requested a totally different folder.
             this._app.open('dir');